docs/reference/NovelAIze-SSR_Source_v3.0/novel_aize_ssr/requirements.txt
aiohttp>=3.9.0
tqdm>=4.66.0
google-genai>=0.3.0
nuitka>=2.0
